+++ docs/meta/history.md
@@ -81,6 +81,10 @@ This is the first release candidate for Hatch v1, a complete rewrite.
 
 ### Unreleased
 
+***Added:***
+
+- Support `text/plain` content type for `project.readme` metadata
+
 ### [0.9.0](https://github.com/ofek/hatch/releases/tag/hatchling-v0.9.0) - 2022-01-18 ### {: #hatchling-v0.9.0 }
 
 ***Added:***